requirements.txt
joblib==1.1.0
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
scikit-learn==0.24.2
scipy==1.6.3
streamlit==1.7.0